+++ WHATSNEW.md
@@ -1,5 +1,14 @@
+Version 3.4 (unreleased)
+========================
+
+* Migration step executors can now throw a `MigrationAbortedException` exception to halt the execution of a migration
+    and have it recorded as either 'done' or 'skipped' instead of 'failed'
+
+* Added a corresponding new event: `ez_migration.migration_aborted` that can be listened to by user code
+
+
 Version 3.3
-=============
+===========
 
 * Fixed: on content creation, assigning a section by identifier instead of id